fix : 빌드오류 - 코드 빠진 것 추가 (#84)
