+++ apps/client/src/features/budget/queries.ts
@@ -1,14 +1,14 @@
-import { budgetClient } from '@/src/shared/api/budget';
+import { BudgetClient } from '@/src/shared/apis/budget';
 import { createQueryKeys } from '@lukemorales/query-key-factory';
 import { Budget } from './types';
 
 export const BudgetKeys = createQueryKeys('budget', {
   getBudget: () => ({
     queryKey: ['get'],
-    queryFn: () => budgetClient.getBudget(),
+    queryFn: () => BudgetClient.getBudget(),
   }),
   updateBudget: (budget: Budget) => ({
     queryKey: [budget],
-    queryFn: () => budgetClient.updateBudget(budget),
+    queryFn: () => BudgetClient.updateBudget(budget),
   }),
 });
