+++ apps/client/src/features/category/queries.ts
@@ -1,9 +1,9 @@
-import { categoryClient } from '@/src/shared/api/category';
+import { CategoryClient } from '@/src/shared/apis/category';
 import { createQueryKeys } from '@lukemorales/query-key-factory';
 
 export const CategoryKeys = createQueryKeys('category', {
   getClassifications: () => ({
     queryKey: ['getClassifications'],
-    queryFn: () => categoryClient.getClassifications(),
+    queryFn: () => CategoryClient.getClassifications(),
   }),
 });
