--- apps/client/src/shared/api/wedding.ts
@@ -1,9 +1,0 @@
-import { axiosInstance } from '../config/axios';
-
-export const weddingClient = {
-  updateTotalBudget: async (totalBudget: number) => {
-    await axiosInstance.put('/totalBudget', {
-      totalBudget,
-    });
-  },
-};
