+++ apps/client/src/shared/apis/category.ts
@@ -4,7 +4,7 @@ import { Pageable } from '@fsd/features/common/types';
 
 const URL_ROOT = '/category';
 
-export const categoryClient = {
+export const CategoryClient = {
   getClassifications: async () => {
     const { data } = await axiosInstance.get<Pageable<Classification>>(URL_ROOT + '/classifications');
     return data;
