+++ apps/client/src/shared/apis/wedding.ts
@@ -1,5 +1,5 @@
-import { axiosInstance } from '../../config/axios';
-import { WeddingInfoType } from '../../types/wedding';
+import { axiosInstance } from '../config/axios';
+import { WeddingInfoType } from '../types/wedding';
 
 const URL_ROOT = '/wedding';
 
@@ -9,5 +9,10 @@ export const WeddingClient = {
   },
   update: async (formData: WeddingInfoType) => {
     return await axiosInstance.put(URL_ROOT, { ...formData });
+  },
+  updateTotalBudget: async (totalBudget: number) => {
+    await axiosInstance.put('/totalBudget', {
+      totalBudget,
+    });
   },
 };
